+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.7.4
+- raid1: Fix degraded discard
+
+## 0.7.3
+- raid1: Bitmap Load/Init
+
+## 0.7.2
+- General fixes
+
 ## 0.7.1
 - raid1: Writes in degraded mode